edited

+++ printf/libft/libft.h
@@ -62,7 +62,7 @@ void	ft_putnbr_fd(int n, int fd);
 
 /* STRING */
 
-int		ft_strlen(const char *i);
+size_t	ft_strlen(const char *i);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
 size_t	ft_strlcat(char *dst, const char *src, size_t size);
 int		ft_toupper(int c);
